Rename class and file
