+++ src/background/actions/action-creator.ts
@@ -11,8 +11,8 @@ import { ScanCompletedPayload } from '../../injected/analyzers/analyzer';
 import { DictionaryNumberTo } from '../../types/common-types';
 import { VisualizationActions } from '../actions/visualization-actions';
 import { VisualizationScanResultActions } from '../actions/visualization-scan-result-actions';
-import { ShortcutsPageController } from '../chrome-shortcuts-page-controller';
 import { DetailsViewController } from '../details-view-controller';
+import { ShortcutsPageController } from '../shortcuts-page-controller';
 import { TargetTabController } from '../target-tab-controller';
 import { TelemetryEventHandler } from '../telemetry/telemetry-event-handler';
 import { ActionHub } from './action-hub';
