+++ src/background/shortcuts-page-controller.ts
