+++ src/tests/unit/tests/background/actions/shortcuts-page-action-creator.test.ts
@@ -5,8 +5,8 @@ import { IMock, It, Mock, Times } from 'typemoq';
 
 import { BaseActionPayload } from '../../../../../background/actions/action-payloads';
 import { ShortcutsPageActionCreator } from '../../../../../background/actions/shortcuts-page-action-creator';
-import { ShortcutsPageController } from '../../../../../background/chrome-shortcuts-page-controller';
 import { Interpreter } from '../../../../../background/interpreter';
+import { ShortcutsPageController } from '../../../../../background/shortcuts-page-controller';
 import { TelemetryEventHandler } from '../../../../../background/telemetry/telemetry-event-handler';
 import { Messages } from '../../../../../common/messages';
 import { SHORTCUT_CONFIGURE_OPEN } from '../../../../../common/telemetry-events';
